src/workshop/settings.py
from pydantic_settings import BaseSettings
from dotenv import load_dotenv
import os

load_dotenv() 
DATABASE_URL = os.getenv('DATABASE_URL')
JWT_SECRET_KEY = os.getenv('JWT_SECRET_KEY')

class Settings(BaseSettings):
    server_host: str = '127.0.0.1'
    server_port: int = 8000
    database_url: str = DATABASE_URL

    jwt_secret: str = JWT_SECRET_KEY
    jwt_algorithm: str = 'HS256'
    jwt_expiration: int = 3600


settings = Settings(
    _env_file='.env',
    _env_file_encoding='utf-8',
)
