requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
sqlalchemy
python-jose
passlib[brcypt]
python-multipart
bcrypt
psycopg2-binary
pytest
pytest-asyncio
